+++ misskey-api/src/model/meta.rs
@@ -1,3 +1,5 @@
+#[cfg(feature = "12-81-0")]
+use crate::model::ad::{Ad, Place};
 #[cfg(feature = "12-62-0")]
 use crate::model::clip::Clip;
 use crate::model::{emoji::Emoji, id::Id, user::User};
@@ -54,6 +56,9 @@ pub struct Meta {
     pub icon_url: Option<String>,
     pub max_note_text_length: u64,
     pub emojis: Vec<Emoji>,
+    #[cfg(feature = "12-81-0")]
+    #[cfg_attr(docsrs, doc(cfg(feature = "12-81-0")))]
+    pub ads: Vec<MetaAd>,
     /// This field is [`bool`] (i.e. not [`Option`]) on <span class="module-item stab portability" style="display: inline-block; font-size: 80%;"><strong>non-<code style="background-color: transparent;">feature="12-58-0"</code></strong></span>.
     #[cfg(feature = "12-58-0")]
     pub require_setup: Option<bool>,
@@ -86,6 +91,18 @@ pub struct Meta {
     #[cfg(feature = "12-62-0")]
     #[cfg_attr(docsrs, doc(cfg(feature = "12-62-0")))]
     pub pinned_clip_id: Option<Id<Clip>>,
+}
+
+#[cfg(feature = "12-81-0")]
+#[cfg_attr(docsrs, doc(cfg(feature = "12-81-0")))]
+#[derive(Deserialize, Serialize, Debug, Clone)]
+#[serde(rename_all = "camelCase")]
+pub struct MetaAd {
+    pub id: Id<Ad>,
+    pub url: String,
+    pub place: Place,
+    pub ratio: u64,
+    pub image_url: String,
 }
 
 #[derive(Deserialize, Serialize, Debug, Clone)]