+++ misskey-util/src/builder/admin.rs
@@ -577,6 +577,8 @@ impl<C> AdBuilder<C> {
             memo: String::default(),
             place: Place::default(),
             priority: Priority::default(),
+            #[cfg(feature = "12-81-0")]
+            ratio: 1,
             expires_at: DateTime::default(),
             image_url: String::default(),
         };
@@ -616,6 +618,13 @@ impl<C> AdBuilder<C> {
         self.place(Place::Horizontal)
     }
 
+    #[cfg(feature = "12-81-0")]
+    #[cfg_attr(docsrs, doc(cfg(feature = "12-81-0")))]
+    /// Sets the place of the ad to horizontal-big.
+    pub fn horizontal_big(&mut self) -> &mut Self {
+        self.place(Place::HorizontalBig)
+    }
+
     /// Sets the priority of the ad.
     pub fn priority(&mut self, priority: Priority) -> &mut Self {
         self.request.priority = priority;
@@ -635,6 +644,14 @@ impl<C> AdBuilder<C> {
     /// Sets the priority of the ad to low.
     pub fn low_priority(&mut self) -> &mut Self {
         self.priority(Priority::Low)
+    }
+
+    #[cfg(feature = "12-81-0")]
+    #[cfg_attr(docsrs, doc(cfg(feature = "12-81-0")))]
+    /// Sets the ratio of the ad.
+    pub fn ratio(&mut self, ratio: u64) -> &mut Self {
+        self.request.ratio = ratio;
+        self
     }
 
     /// Sets the expiration date of the ad.
@@ -682,6 +699,8 @@ impl<C> AdUpdateBuilder<C> {
             expires_at,
             place,
             priority,
+            #[cfg(feature = "12-81-0")]
+            ratio,
             url,
             image_url,
             memo,
@@ -693,6 +712,8 @@ impl<C> AdUpdateBuilder<C> {
             memo,
             place,
             priority,
+            #[cfg(feature = "12-81-0")]
+            ratio,
             expires_at,
             image_url,
         };
@@ -732,6 +753,13 @@ impl<C> AdUpdateBuilder<C> {
         self.place(Place::Horizontal)
     }
 
+    #[cfg(feature = "12-81-0")]
+    #[cfg_attr(docsrs, doc(cfg(feature = "12-81-0")))]
+    /// Sets the place of the ad to horizontal-big.
+    pub fn horizontal_big(&mut self) -> &mut Self {
+        self.place(Place::HorizontalBig)
+    }
+
     /// Sets the priority of the ad.
     pub fn priority(&mut self, priority: Priority) -> &mut Self {
         self.request.priority = priority;
@@ -751,6 +779,14 @@ impl<C> AdUpdateBuilder<C> {
     /// Sets the priority of the ad to low.
     pub fn low_priority(&mut self) -> &mut Self {
         self.priority(Priority::Low)
+    }
+
+    #[cfg(feature = "12-81-0")]
+    #[cfg_attr(docsrs, doc(cfg(feature = "12-81-0")))]
+    /// Sets the ratio of the ad.
+    pub fn ratio(&mut self, ratio: u64) -> &mut Self {
+        self.request.ratio = ratio;
+        self
     }
 
     /// Sets the expiration date of the ad.